cheapskatecoderbackend/requirements.txt
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.2.1
django-extensions==2.2.9
django-rest-knox==4.1.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
ipython==7.14.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
